counter测试
