+++ src-client/pages/admin/containers/PageTestCounter.js
@@ -1,16 +1,52 @@
 import React, {PropTypes, Component} from 'react';
+import {connect} from 'react-redux'
 
-class PageUIDemoTest1 extends Component {
+import {increase, decrease} from '../actions/count'
+
+class PageTestCounter extends Component {
+  static propTypes = {
+    number: PropTypes.number.isRequired,
+    increase: PropTypes.func.isRequired,
+    decrease: PropTypes.func.isRequired
+  }
+
   constructor(props, context) {
     super(props, context);
-    console.log('==test/counter== constructor');
+
+    this.handleIncreaseClick = this.handleIncreaseClick.bind(this);
+    this.handleDecreaseClick = this.handleDecreaseClick.bind(this);
+  }
+
+  handleIncreaseClick() {
+    this.props.increase(1);
+  }
+
+  handleDecreaseClick() {
+    this.props.decrease(1);
   }
 
   render() {
+    let {number} = this.props;
+
     return (
-      <div className="fk-content-wrap">hello test/counter</div>
-    )
+      <div className="fk-content-wrap">
+        <h2>HELLO, index</h2>
+        <div>
+          Some state changes:
+          {number}
+          <br/><br/><br/>
+          <button onClick={this.handleIncreaseClick}>Increase</button>
+          <button onClick={this.handleDecreaseClick}>Decrease</button>
+        </div>
+      </div>
+    );
   }
 }
 
-export default  PageUIDemoTest1;
+const mapStateToProps = (state) => ({
+  number: state.count.number
+});
+
+const mapDispatchToProps = {increase, decrease};
+
+export default connect(mapStateToProps, mapDispatchToProps)(PageTestCounter);
